Generalized add/update form
